@@ -7,7 +7,7 @@ import 'package:intl/intl.dart';
 import 'dart:convert';
 
 class CountryNames {
-  static CountryNames of(BuildContext context) {
+  static CountryNames? of(BuildContext context) {
     return Localizations.of<CountryNames>(context, CountryNames);
   }
 
@@ -15,7 +15,7 @@ class CountryNames {
   final Map<String, String> data;
   CountryNames(this.locale, this.data);
 
-  String nameOf(String code) => data[code];
+  String? nameOf(String code) => data[code];
 
   List<MapEntry<String, String>> get sortedByCode {
     return data.entries.toList()..sort((a, b) => a.key.compareTo(b.key));
@@ -26,12 +26,14 @@ class CountryNames {
   }
 }
 
-class CountryNamesLocalizationsDelegate extends LocalizationsDelegate<CountryNames> {
-  final AssetBundle bundle;
+class CountryNamesLocalizationsDelegate
+    extends LocalizationsDelegate<CountryNames> {
+  final AssetBundle? bundle;
   const CountryNamesLocalizationsDelegate({this.bundle});
 
   Future<List<String>> locales() async {
-    return List<String>.from(await _loadJSON('languages.json') as List<dynamic>);
+    return List<String>.from(
+        await _loadJSON('languages.json') as List<dynamic>);
   }
 
   @override
@@ -39,21 +41,22 @@ class CountryNamesLocalizationsDelegate extends LocalizationsDelegate<CountryNam
 
   @override
   Future<CountryNames> load(Locale locale) async {
-    final String name = locale.countryCode == null ? locale.languageCode : locale.toString();
+    final String name =
+        locale.countryCode == null ? locale.languageCode : locale.toString();
     final String localeName = Intl.canonicalizedLocale(name);
 
     var locales = Set<String>.from(await this.locales());
 
     var availableLocale = Intl.verifiedLocale(
-      localeName,
-      (locale) => locales.contains(locale),
-      onFailure: (_) => 'en'
-    );
+        localeName, (locale) => locales.contains(locale),
+        onFailure: (_) => 'en');
+
     if (availableLocale == null) {
-      return null;
+      throw Exception('Locale not found');
     }
 
-    final data = Map<String, String>.from(await _loadJSON('data/$availableLocale.json') as Map<dynamic, dynamic>);
+    final data = Map<String, String>.from(
+        await _loadJSON('data/$availableLocale.json') as Map<dynamic, dynamic>);
     return CountryNames(availableLocale, data);
   }
 
@@ -65,6 +68,7 @@ class CountryNamesLocalizationsDelegate extends LocalizationsDelegate<CountryNam
   Future<dynamic> _loadJSON(key) {
     Future<dynamic> parser(String data) async => jsonDecode(data);
     final bundle = this.bundle ?? rootBundle;
-    return bundle.loadStructuredData('packages/flutter_localized_countries/' + key, parser);
+    return bundle.loadStructuredData(
+        'packages/flutter_localized_countries/' + key, parser);
   }
 }
