@@ -10,13 +10,16 @@ class TestAssetBundle extends CachingAssetBundle {
   @override
   Future<ByteData> load(String key) async {
     const prefix = "packages/flutter_localized_countries/";
+
     if (key.startsWith(prefix)) {
-      var path = join(dirname(Platform.script.toFilePath()), key.substring(prefix.length));
+      var path = join(
+          dirname(Platform.script.toFilePath()), key.substring(prefix.length));
       var bytes = Uint8List.fromList(await File(path).readAsBytes());
       var buffer = bytes.buffer;
       return ByteData.view(buffer);
     }
-    return null;
+
+    return throw Exception();
   }
 }
 
@@ -33,7 +36,7 @@ void main() {
     var matcher = completion(predicate(f, 'name of the $cc is "$name"'));
     expect(d.load(locale), matcher);
   }
-  
+
   test('localizes by language', () {
     checkTranslation(Locale('de'), 'CH', 'Schweiz');
     checkTranslation(Locale('en'), 'CH', 'Switzerland');
